backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pymupdf>=1.24.0
pdfplumber>=0.11.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
spacy>=3.7.0
watchdog>=4.0.0
python-multipart>=0.0.9
websockets>=12.0
google-generativeai>=0.7.0
numpy>=1.26.0
